optional images-4

+++ src/car-image/car-image.service.ts
@@ -9,10 +9,36 @@ import { isNullValue } from "../helper/utils";
 export class CarImageService {
   constructor(private readonly prisma: PrismaService) {
   }
-  create(createCarImageDto: CreateCarImageDto[]) {
-    return this.prisma.carImage.createMany({
+  async create(createCarImageDto: CreateCarImageDto[]) {
+    let res={};
+    await this.prisma.carImage.deleteMany({
+      where: {
+        AND: [
+          {
+            carId: createCarImageDto[0].carId
+          },
+          {
+            OR: [
+              {
+                url: 'car_image_1.png'
+              },
+              {
+                url: 'car_image_2.png'
+              },
+              {
+                url: 'car_image_3.png'
+              }
+            ]
+          }
+        ]
+      }
+    })
+    await this.prisma.carImage.createMany({
       data: createCarImageDto
-    });
+    }).then(result=>{
+      res = result;
+    })
+    return res;
   }
 
   async update(id: number, updateCarImageDto: CreateCarImageDto) {
@@ -40,7 +66,10 @@ export class CarImageService {
       }
     })
     if(!isNullValue(oldData)){
-      await fs.unlink(`./upload/car/image/${oldData.url}`,()=>{});
+      if(oldData.url==="car_image_1.png" || oldData.url==="car_image_2.png" || oldData.url==='car_image_3.png'){}
+      else {
+        await fs.unlink(`./upload/car/image/${oldData.url}`,()=>{});
+      }
     }
     return this.prisma.carImage.delete({
       where: {
